@@ -1,26 +1,24 @@
 <?php
 
-namespace Borsaco\TelegramBotApiBundle\Command;
+namespace Oihso\TelegramBotApiBundle\Command;
 
-use Borsaco\TelegramBotApiBundle\Service\Bot;
+use Oihso\TelegramBotApiBundle\Service\Bot;
+use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ChoiceQuestion;
-use Symfony\Component\Console\Question\Question;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Telegram\Bot\Exceptions\TelegramSDKException;
-use TelegramBot\Api\BotApi;
 
+#[AsCommand(name: 'telegram:bot:webhook:info')]
 class WebhookInfoCommand extends Command
 {
     /**
      * @var Bot
      */
-    private $bot;
-
-    protected static $defaultName = 'telegram:bot:webhook:info';
+    private Bot $bot;
 
     /**
      * @inheritDoc
@@ -35,8 +33,8 @@ class WebhookInfoCommand extends Command
     /**
      * @inheritDoc
      */
-    protected function configure()
-    {
+    protected function configure(): void
+	{
         $this
             ->addArgument('name', InputArgument::OPTIONAL, 'The bot name (is set in configuration file)')
             ->setDescription('Webhook Information')
