@@ -1,8 +1,7 @@
 <?php
 
-namespace Borsaco\TelegramBotApiBundle\DependencyInjection;
+namespace Oihso\TelegramBotApiBundle\DependencyInjection;
 
-use App\Kernel;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
