@@ -1,14 +1,18 @@
 <?php
 
-namespace Borsaco\TelegramBotApiBundle\DependencyInjection\Factory;
+namespace Oihso\TelegramBotApiBundle\DependencyInjection\Factory;
 
 use GuzzleHttp\Client;
 use Telegram\Bot\Api;
+use Telegram\Bot\Exceptions\TelegramSDKException;
 use Telegram\Bot\HttpClients\GuzzleHttpClient;
 
 class BotFactory
 {
-    public function create(array $config, string $name): Api
+	/**
+	 * @throws TelegramSDKException
+	 */
+	public function create(array $config, string $name): Api
     {
         $bot = new Api($config['bots'][$name]['token']);
 
