@@ -1,7 +1,8 @@
 <?php
 
-namespace Borsaco\TelegramBotApiBundle\DependencyInjection;
+namespace Oihso\TelegramBotApiBundle\DependencyInjection;
 
+use Exception;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
@@ -9,8 +10,11 @@ use Symfony\Component\DependencyInjection\Loader;
 
 class TelegramBotApiExtension extends Extension
 {
-    public function load(array $configs, ContainerBuilder $container)
-    {
+	/**
+	 * @throws Exception
+	 */
+	public function load(array $configs, ContainerBuilder $container): void
+	{
         $configuration = new Configuration();
         $config = $this->processConfiguration($configuration, $configs);
 
